+++ lib/pg_search/features/tsearch.rb
@@ -4,6 +4,10 @@ require "active_support/core_ext/module/delegation"
 module PgSearch
   module Features
     class TSearch < Feature
+      def self.valid_options
+        super + [:dictionary, :prefix, :negation, :any_word, :normalization, :tsvector_column]
+      end
+
       def initialize(*args)
         super
 
@@ -28,8 +32,7 @@ module PgSearch
 
       DISALLOWED_TSQUERY_CHARACTERS = /['?\\:]/
 
-      def tsquery_for_term(unsanitized_term)
-
+      def tsquery_for_term(unsanitized_term) # rubocop:disable Metrics/AbcSize
         if options[:negation] && unsanitized_term.start_with?("!")
           unsanitized_term[0] = ''
           negated = true