+++ lib/pg_search/multisearchable.rb
@@ -2,7 +2,7 @@ require "active_support/core_ext/class/attribute"
 
 module PgSearch
   module Multisearchable
-    def self.included mod
+    def self.included(mod)
       mod.class_eval do
         has_one :pg_search_document,
           :as => :searchable,
@@ -10,11 +10,17 @@ module PgSearch
           :dependent => :delete
 
         after_save :update_pg_search_document,
-          :if => lambda { PgSearch.multisearch_enabled? }
+          :if => -> { PgSearch.multisearch_enabled? }
       end
     end
 
-    def update_pg_search_document
+    def searchable_text
+      Array(pg_search_multisearchable_options[:against])
+        .map { |symbol| send(symbol) }
+        .join(" ")
+    end
+
+    def update_pg_search_document # rubocop:disable Metrics/AbcSize
       if_conditions = Array(pg_search_multisearchable_options[:if])
       unless_conditions = Array(pg_search_multisearchable_options[:unless])
 
@@ -23,7 +29,8 @@ module PgSearch
         unless_conditions.all? { |condition| !condition.to_proc.call(self) }
 
       if should_have_document
-        pg_search_document ? pg_search_document.save : create_pg_search_document
+        (pg_search_document || build_pg_search_document)
+          .update_attributes(content: searchable_text)
       else
         pg_search_document.destroy if pg_search_document
       end