+++ spec/spec_helper.rb
@@ -1,3 +1,6 @@
+require "codeclimate-test-reporter"
+CodeClimate::TestReporter.start
+
 require "bundler/setup"
 require "pg_search"
 
@@ -9,13 +12,15 @@ RSpec.configure do |config|
   config.mock_with :rspec do |c|
     c.syntax = :expect
   end
+
+  config.example_status_persistence_file_path = 'tmp/examples.txt'
 end
 
 require 'support/database'
 require 'support/with_model'
 
 DOCUMENTS_SCHEMA = lambda do |t|
-  t.belongs_to :searchable, :polymorphic => true
+  t.belongs_to :searchable, :polymorphic => true, :index => true
   t.text :content
   t.timestamps null: false
 end